requirements.txt
Django>=5.2.3
djangorestframework
djangorestframework-simplejwt
django-filter
pillow
